+++ CHANGELOG.md
@@ -23,6 +23,7 @@ listed in the changelog.
 
 - Release branches of subrepos are not detected ([#269](https://github.com/opendevstack/ods-pipeline/pull/269))
 - `Directory` values in the artifact manifest (`.ods/artifacts/manifest.json`) contain an erronous leading slash. This should only be an issue if you relied on this value in a custom task. ([#269](https://github.com/opendevstack/ods-pipeline/pull/269))
+- `ods-finish` does not upload artifacts of subrepos ([#257](https://github.com/opendevstack/ods-pipeline/issues/257))
 
 ## [0.1.1] - 2021-10-28
 ### Fixed